dohs and plot model updated

+++ src/Model/Table/PlotsTable.php
@@ -58,7 +58,7 @@ class PlotsTable extends Table
     public function validationDefault(Validator $validator)
     {
         $validator
-            ->add('id', 'valid', ['rule' => 'integer'])
+            ->add('id', 'valid', ['rule' => 'numeric'])
             ->allowEmpty('id', 'create');
             
         $validator
